test: type the tile-mapped base address as uintptr_t

+++ test/npu-xrt/tile_mapped_read/kernel.cpp
@@ -9,7 +9,7 @@
 
 extern "C" {
 
-constexpr uint32_t tm_start_addr = 0x80000;
+constexpr uintptr_t tm_start_addr = 0x80000;
 
 void read_processor_bus(uint32_t *data, uint32_t addr, uint32_t size,
                         uint32_t stride) {
